refactored get vehicle make and model to tech records service
